[bug]fix compile error in js_thread.cpp
description: fix compile error in js_thread.cpp

reason: The reason is comparison of integers of different signs 'size_t' and 'int'. So change 'size_t' to 'int'.

issue: https://gitee.com/openharmony/ark_js_runtime/issues/I4YVD3

+++ ecmascript/js_thread.cpp
@@ -284,7 +284,7 @@ void JSThread::LoadStubModule(const char *moduleFile)
         && asmInterOpt.handleEnd >= 0 && asmInterOpt.handleEnd < kungfu::BytecodeStubCSigns::NUM_OF_ALL_STUBS
         && asmInterOpt.handleStart <= asmInterOpt.handleEnd) {
         for (int i = asmInterOpt.handleStart; i <= asmInterOpt.handleEnd; i++) {
-            glueData_.bcHandlers_.Set(i, defaultBCHandlerDes.codeAddr_);
+            glueData_.bcHandlers_.Set(static_cast<size_t>(i), defaultBCHandlerDes.codeAddr_);
         }
     }
 #ifdef NDEBUG
